Refactored dropdown search bar
* Moved the dropdown component to its own folder
* Refactored dropdown code into function rather than component
* Moved images to assets folder
