+++ src/components/dropdown-search/index.css
@@ -1,0 +1,18 @@
+.dropdown {
+	display: inline-block;
+  width: 100%;
+  padding: 10px;
+  background-color: black;
+  text-align: center;
+  border-bottom: solid 5px #969696;
+  position: fixed;
+  z-index: 1;
+  transition: margin-top 0.3s ease;
+  background-image: url("../../assets/background.png");
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+.dropdown.hidden {
+  margin-top: -69px;
+}
