+++ src/main.scss
@@ -4,7 +4,7 @@
   -moz-box-sizing: border-box;
 }
 @mixin background {
-  background-image: url('./images/background.png');
+  background-image: url('./assets/background.png');
   background-repeat: no-repeat;
   background-size: cover;
 }
@@ -104,7 +104,7 @@ select {
   font-family: 'Bitter', serif;
   font-size: 20px;
   background-color: white;
-  background-image: url('./images/downarrow.png');
+  background-image: url('./assets/downarrow.png');
   background-position: calc(100% - 10px);
   background-repeat: no-repeat;
   background-size: 20px 12.50px;
@@ -115,24 +115,6 @@ select {
     outline: none;
     border: solid 5px $darkColor;
   }
-}
-
-.dropdown {
-	width: 100%;
-	padding: 10px;
-	background-color: black;
-	text-align: center;
-	border-bottom: solid 5px $darkColor;
-	display: inline-block;
-	position: fixed;
-	z-index: 1;
-	transition: margin-top 0.3s ease;
-
-	@include background
-
-	&.hidden {
-		margin-top: -69px;
-	}
 }
 
 .header-body {
